add error handling for posting comment
